some cleaning up
a few modifications, mostly just me working through the code to figure
out what's going on and why some decisions were made.

'file' parameter to read_hoof_files() exists only because Matlab's dir()
function really wants to return '.' and '..' -- there may be a less
inclusive version of the function that will ignore dotfiles -- need to
look to see if that's a possibility.