dropSet fixed to actually represent all the files
i'm a bit of an idiot and didn't realize that i should be making those
data arrays. i'm pretty sure it was just loading the last file's data as
of last commit.

frankly, there should probably be an additional "drop" object -- that
would resolve a lot of the organizational issues in the class and the
fact that the arrays of data are a bit clunky.