implemented basic maximum calculations in drop
not doing any windowing of data (so not accounting for any second bounces
or anything like that).

currently taking the absolute value of the xyz accelerations but not the
xyz loads. that may be the wrong decision -- need to look at the data
output more closely to figure that out.