absolute value of loadxyz & comment fixes
it looks like rectifying the signal on the loads is the right thing to do.
doesn't have an effect with the test data set for x or y but max is
significantly higher for z if rectified.

also checked -- in sample data set rectifying the accelerations has no
effect on their max_values