nothing works yet.
moving all of the process functions into their own files. that seems like
the best way to do things for right now at least. i'll add the directory
to the matlab path and  have a separate setup script to create function
handles for all of the functions in it. once the gui is more robust, that
action will take place within the gui setup.

we'll have a method in dropSet that will basically get passed a function
handle/a set of function handles and will apply them to the dropSet. we're
basically mapping the function to the data (think haskell)

don't know how output data is going to be handled in the future. for right
now it'll stay in the drop class' parameters, but obviously that won't
work with the modularity idea in the future because to add a process the
user would have to modify drop and create new output parameters as well as
just writing the function and dropping it in out processes directory.