refactored processing on data
all processing now done through user processes in the processes/ directory and the processor class. computed values are also stored in a processor object -- unless they are calibrated versions of signals from drops, in which case the drop data itself is modified and nothing is stored in the processor object.

also contains an initializer script that sets up a few things for the development environment -- this will be replaced by functionality in the gui setup scripts once the gui is ready for integration into master