refactor and feature addition for sweden compat
matDrop and asciiDrop now represent the two cases for drop data format.
both are subclasses of the abstract class drop.

additionally, dropSet was restructured to reflect that there are now two
possible routes for data read.

sample rate needs some looking after -- given that it is not actually a
constant as it may be different between machines.

need to confirm which channel is which for xyz -- load and acc for US
and UK machines.

going to have to add some tests to see if triaxial vs no triaxial load
cell is handled properly for both ascii and mat data formats.