Allow column offset to handle date columns
modifies expectations of input data files so that we can ignore columns on the left that are non-numeric in addition to ignoring header lines.

Longer term we should be trying to allow for arbitrary data types in arbitrary columns (so that we don't have to require the rectangular table left after ignoring headerlines and columns to contain only numeric values.